% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_sf.R
\name{block_sf}
\alias{block_sf}
\title{block_sf}
\usage{
block_sf(yr)
}
\arguments{
\item{yr}{The vintage year for the geography (2010 or 2020)}
}
\value{
An sf object projected into EPSG: 26986 - NAD83 / Massachusetts Mainland State Plane (meters)
}
\description{
Downloads a shapefile / sf object of Census Blocks from the tigris package
}
\examples{
#pending
}
