% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockgroup_sf.R
\name{blockgroup_sf}
\alias{blockgroup_sf}
\title{blockgroup_sf}
\usage{
blockgroup_sf(yr)
}
\arguments{
\item{yr}{The vintage year for the geography (2010 or 2020)}
}
\value{
An sf object projected into EPSG: 26986 - NAD83 / Massachusetts Mainland State Plane (meters)
}
\description{
Downloads a shapefile / sf object of Census Block Groups from the tigris package
}
\examples{
#pending
}
