% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_xw_bg.R
\docType{data}
\name{census_xw_bg}
\alias{census_xw_bg}
\title{Weighted crosswalk for 2020 Census Block Group values in 2010 Block Group boundaries}
\format{
A tibble with 4985 rows and 10 variables:
\describe{
\item{bg10_id}{dbl 2010 Census Block Group FIPS Code}
\item{hu_2020}{dbl 2020 Housing Unit counts in 2010 Block Group geography}
\item{pop_2020}{dbl 2020 Population counts in 2010 Block Group geography}
\item{ohu_2020}{dbl 2020 Occupied Housing Unit counts in 2010 Block Group geography}
\item{hu_2010}{dbl 2010 Housing Unit counts in 2010 Block Group geography}
\item{pop_2010}{dbl 2010 Population counts in 2010 Block Group geography}
\item{hu_chng}{dbl Change in housing unit counts from 2010 to 2020}
\item{hu_chng_p}{dbl Percent change in housing unit counts from 2010 to 2020}
\item{pop_chng}{dbl Change in population from 2010 to 2020}
\item{pop_chng_p}{dbl Percent change in population from 2010 to 2020}
}
}
\usage{
census_xw_bg
}
\description{
This crosswalk provides 2030 Block Group level data on Housing Units and Population in 2010 census Block Group
geographies for easy cross-comparison.
}
\keyword{datasets}
