% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_xw_bg20.R
\docType{data}
\name{census_xw_bg20}
\alias{census_xw_bg20}
\title{Weighted crosswalk for 2010 & 2020 Census Block Group values in 2020 Block Group boundaries}
\format{
A tibble with 5,116 rows and 9 variables:
\describe{
\item{bg20_id}{dbl 2010 Census Block Group FIPS Code}
\item{hu_2020}{dbl 2020 Housing Unit counts in 2010 block geography}
\item{pop_2020}{dbl 2020 Population counts in 2010 block geography}
\item{hu_2010}{dbl 2010 Housing Unit counts in 2010 block geography}
\item{pop_2010}{dbl 2010 Population counts in 2010 block geography}
\item{huch1020}{dbl Change in housing unit counts from 2010 to 2020}
\item{huch1020p}{dbl Percent change in housing unit counts from 2010 to 2020}
\item{popch1020}{dbl Change in population from 2010 to 2020}
\item{popch1020p}{dbl Percent change in population from 2010 to 2020}
}
}
\usage{
census_xw_bg20
}
\description{
This crosswalk provides 2020 Block Group level data on Housing Units and Population in 2020 census Block Group
geographies for easy cross-comparison.
}
\keyword{datasets}
