% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_geog.R
\name{crosswalk_geog}
\alias{crosswalk_geog}
\title{Function to call desired nested geographies for blocks for both 2010 and 2020}
\usage{
crosswalk_geog(year)
}
\arguments{
\item{year}{Desired Census Year}
}
\value{
Call nested geographies for Census Blocks
}
\description{
Function to call desired nested geographies for blocks for both 2010 and 2020
}
\examples{

crosswalk_geog(year = 2010)
}
