% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_wtd.R
\name{crosswalk_wtd}
\alias{crosswalk_wtd}
\title{Function to call desired weighted crosswalk for 2020 values in 2010 census geographies}
\usage{
crosswalk_wtd(geography, vintage = NULL)
}
\arguments{
\item{geography}{Desired census geography. Can be either 'block', 'block group', 'tract' or 'county subdivision'}

\item{vintage}{Desired Census Year (either 2010, or 2020)}
}
\value{
Geographically weighted values for housing units and population by desired vintage and unit of geography
}
\description{
Function to call desired weighted crosswalk for 2020 values in 2010 census geographies
}
\examples{

crosswalk_wtd(geography = 'block', vintage = 2010)
}
