% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_xw_2020.R
\docType{data}
\name{geog_xw_2020}
\alias{geog_xw_2020}
\title{Crosswalk Between 2020 Census blocks with 2020 tracts, 2020 block groups, and MA municipalities & counties}
\format{
A tibble with 107202  rows and 10 variables:
\describe{
\item{bl20_id}{chr 2020 Census Block FIPS Code}
\item{bg20_id}{chr 2020 Census Block Group FIPS Code}
\item{ct20_id}{chr 2020 Census Tract FIPS Code}
\item{cosub_cn20}{chr 2020 County Subdivision FIPS Code}
\item{muni_id}{dbl Municipal ID Number 1 Abington through 351 Yarmouth}
\item{muni_name}{chr Official Municipality name}
\item{county}{chr Massachusetts County Name}
\item{statefp}{chr Massachusetts state FIPS Code}
\item{countyfp}{chr Massachusetts county FIPS Code}
\item{cousubfp}{chr Massachusetts county subdivision FIPS Code}
}
}
\usage{
geog_xw_2020
}
\description{
This crosswalk provides an easy way to categorise the 2020 blocks into various geographies
}
\keyword{datasets}
