% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapc_data_keys.R
\docType{data}
\name{mapc_data_keys}
\alias{mapc_data_keys}
\title{Metropolitan Area Planning Council Information}
\format{
A tibble with 351 rows and 9 variables:
\describe{
\item{muni_id}{dbl Municipal ID Number 1 Abington through 351 Yarmouth}
\item{muni_name}{chr Official Municipality name}
\item{mapc}{dbl Binary indicating MAPC affiliation}
\item{subrg_id}{dbl MAPC Subregion ID}
\item{subrg_nm}{chr MAPC Subregion Name}
\item{subrg_acr}{chr MAPC Subregion Acronym}
\item{subrg_alt}{chr Secondary MAPC Subregion}
\item{mmc}{dbl Whether part of the Metro Mayors' Coalition}
\item{nsc}{dbl Whether part of the North Shore Coalition}
}
}
\usage{
mapc_data_keys
}
\description{
These data keys represent the MAPC affiliation for the 351 municipalities in Massachusetts.
Other indicators such as MAPC Sub Regions, and coalitions such as the Metro Mayors' Coalition, and the North Shore Coalition are also indicated.
A few are members of more than one Sub Region, and thus the 'subrg_alt' column indicates the secondary Sub Region in such cases.
}
\keyword{datasets}
