% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muni_sf.R
\name{muni_sf}
\alias{muni_sf}
\title{muni_sf}
\usage{
muni_sf(yr)
}
\arguments{
\item{yr}{The vintage year for the geography (2010 or 2020)}
}
\value{
An sf object projected into EPSG: 26986 - NAD83 / Massachusetts Mainland State Plane (meters)
}
\description{
Downloads a shapefile / sf object of Massachusetts municipal boundaries from the tigris package
}
\examples{
#pending
}
