% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puma_muni_xw_2020.R
\docType{data}
\name{puma_muni_xw_2020}
\alias{puma_muni_xw_2020}
\title{Crosswalk Between 2020 PUMAs and all municipalities or portions of municipalities contained within}
\format{
A tibble with 463 rows and 9 variables:
\describe{
\item{PUMACE20}{chr 2020 PUMA Code}
\item{muni_id}{dbl Municipal ID Number 1 Abington through 351 Yarmouth}
\item{muni_name}{chr Official Municipality name}
\item{cosub_cn20}{chr 2020 Census GEOID for county subdivision (muni)}
\item{pct_muni_area_overlap}{num Percent of municipality's total area that overlaps with PUMA}
\item{muni_area_m2}{num Total area in square meters of municipality}
\item{puma20_area_m2}{num Total area in square meters of 2020 PUMA}
\item{join_type}{chr method of spatial join applied to the two layers}
}
}
\usage{
puma_muni_xw_2020
}
\description{
This crosswalk provides a way to identify which municipalities (or portions thereof) lay within each of the 2020 PUMAs
}
\keyword{datasets}
