% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puma_sf.R
\name{puma_sf}
\alias{puma_sf}
\title{puma_sf}
\usage{
puma_sf(yr)
}
\arguments{
\item{yr}{The vintage year for the geography (2010 or 2020).}
}
\value{
An sf object projected into EPSG: 26986 - NAD83 / Massachusetts Mainland State Plane (meters)
}
\description{
Downloads a shapefile / sf object of Census Public Use Microdata Areas (PUMAs) from the tigris package.
}
\details{
New PUMA geographies are typically created 2-3 years after each decennial census.

Census 5-year ACS and PUMS datasets will contain both the "old" and "new" PUMA codes when the 5-year period
spans the introduction year of the new PUMA geographies.

Selecting yr = 2020 for this function will produce the 2020 vintage PUMAs that first show up in the 2022
ACS/PUMS datasets. In mixed-vintage PUMS datasets, there will be two different ID codes for the different PUMA
vintages: PUMACE10 for the 2010 vintage and PUMACE20 for the 2020 vintage, depending on when the household was
surveyed in the 5-year period.

From 2023 onwards, the Census Bureau has recoded houesholds into the new 2020
PUMA geographies, so crosswalking is not required.
}
\examples{
#pending
}
