% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpa_data_keys.R
\docType{data}
\name{rpa_data_keys}
\alias{rpa_data_keys}
\title{Regional Planning Authorities of Massachusetts}
\format{
A tibble with 351 rows and 6 variables:
\describe{
\item{muni_id}{dbl Municipal ID Number 1 Abington through 351 Yarmouth}
\item{muni_name}{chr Official Municipality name}
\item{rpa_id}{dbl Regional Planning Authority ID}
\item{rpa_name}{chr Regional Planning Authority Name}
\item{rpa_acr}{chr Regional Planning Authority Acronym}
\item{rpa_alt}{chr Secondary Regional Planning Authority}
}
}
\usage{
rpa_data_keys
}
\description{
These data keys represent the different RPA affiliations for the 351 municipalities in Massachusetts.
A few are members of more than one RPA, and thus the 'rpa_alt' column indicates the secondary RPA in such cases.
}
\keyword{datasets}
