% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tract_sf.R
\name{tract_sf}
\alias{tract_sf}
\title{Downloads a shapefile / sf object of Census Tracts from the tigris package}
\usage{
tract_sf(yr)
}
\arguments{
\item{yr}{The vintage year for the geography (2010 or 2020)}
}
\value{
An sf object projected into EPSG: 26986 - NAD83 / Massachusetts Mainland State Plane (meters)
}
\description{
Downloads a shapefile / sf object of Census Tracts from the tigris package
}
\examples{
#pending
}
