% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapcdatakeys-package.R
\docType{package}
\name{mapcdatakeys-package}
\alias{mapcdatakeys}
\alias{mapcdatakeys-package}
\title{mapcdatakeys: Provides Various Data Keys for Various Massachusetts Geographies}
\description{
This package provides various data keys for different 
    Massachusetts political and census geographies, 
    such as municipalities, census tracts, counties, etc.
}
\author{
\strong{Maintainer}: MAPC Boston, MA \email{adeodhar@mapc.org} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
